Switched the position of the gallery and the close button
