+++ packages/inspection-capture-web/src/PhotoCapture/PhotoCaptureHUD/PhotoCaptureHUDElementsSight/PhotoCaptureHUDElementsSight.tsx
@@ -1,55 +1,10 @@
-import { Image, PixelDimensions, Sight } from '@monkvision/types';
 import { SightOverlay } from '@monkvision/common-ui-web';
 import { SightSlider } from './SightSlider';
 import { styles } from './PhotoCaptureHUDElementsSight.styles';
 import { AddDamageButton } from './AddDamageButton';
-import { usePhotoCaptureHUDSightPreviewStyle } from './hooks';
+import { PhotoCaptureHUDElementsSightProps, usePhotoCaptureHUDSightPreviewStyle } from './hooks';
 import { PhotoCaptureHUDCounter } from '../PhotoCaptureHUDCounter';
 import { PhotoCaptureMode } from '../../hooks';
-
-/**
- * Props of the PhotoCaptureHUDElementsSight component.
- */
-export interface PhotoCaptureHUDElementsSightProps {
-  /**
-   * The list of sights provided to the PhotoCapture component.
-   */
-  sights: Sight[];
-  /**
-   * The currently selected sight in the PhotoCapture component : the sight that the user needs to capture.
-   */
-  selectedSight: Sight;
-  /**
-   * Callback called when the user manually select a new sight.
-   */
-  onSelectedSight?: (sight: Sight) => void;
-  /**
-   * Callback called when the user manually select a sight to retake.
-   */
-  onRetakeSight?: (sight: string) => void;
-  /**
-   * Callback called when the user clicks on the AddDamage button.
-   */
-  onAddDamage?: () => void;
-  /**
-   * Array containing the list of sights that the user has already captured.
-   */
-  sightsTaken: Sight[];
-  /**
-   * The dimensions of the Camera video stream.
-   */
-  streamDimensions?: PixelDimensions | null;
-  /**
-   * The current images taken by the user (ignoring retaken pictures etc.).
-   */
-  images: Image[];
-  /**
-   * Boolean indicating whether the Add Damage feature is disabled. If disabled, the `Add Damage` button will be hidden.
-   *
-   * @default true
-   */
-  enableAddDamage?: boolean;
-}
 
 /**
  * Component implementing an HUD displayed on top of the Camera preview during the PhotoCapture process when the current
@@ -66,7 +21,7 @@ export function PhotoCaptureHUDElementsSight({
   images,
   enableAddDamage,
 }: PhotoCaptureHUDElementsSightProps) {
-  const style = usePhotoCaptureHUDSightPreviewStyle(streamDimensions);
+  const style = usePhotoCaptureHUDSightPreviewStyle({ streamDimensions });
 
   return (
     <div style={styles['container']}>
