+++ packages/inspection-capture-web/src/PhotoCapture/PhotoCaptureHUD/PhotoCaptureHUDElementsSight/index.ts
@@ -1,6 +1,4 @@
 export * from './SightSlider';
 export * from './AddDamageButton';
-export {
-  PhotoCaptureHUDElementsSight,
-  type PhotoCaptureHUDElementsSightProps,
-} from './PhotoCaptureHUDElementsSight';
+export { PhotoCaptureHUDElementsSight } from './PhotoCaptureHUDElementsSight';
+export { type PhotoCaptureHUDElementsSightProps } from './hooks';
